+++ Tests/YamsTests/YamlErrorTests.swift
@@ -149,6 +149,52 @@ class YamlErrorTests: XCTestCase {
             )
         }
     }
+
+    func testDuplicateKeysCannotBeParsed() throws {
+        let yamlString = """
+            a: value
+            a: different_value
+        """
+        XCTAssertThrowsError(try Parser(yaml: yamlString).singleRoot()) { error in
+            XCTAssertTrue(error is YamlError)
+            XCTAssertEqual("\(error)", """
+                error: parser: expected all keys to be unique but found the following duplicated key(s):
+                a ([1:5, 2:5]):
+                    a: value
+                    ^
+                    a: different_value
+                    ^
+                """)
+        }
+    }
+
+    func testDuplicatedKeysCannotBeParsed_MultipleDuplicates() throws {
+        let yamlString = """
+            a: value
+            a: different_value
+            b: value
+            b: different_value
+            b: different_different_value
+        """
+        XCTAssertThrowsError(try Parser(yaml: yamlString).singleRoot()) { error in
+            XCTAssertTrue(error is YamlError)
+            XCTAssertEqual("\(error)", """
+                error: parser: expected all keys to be unique but found the following duplicated key(s):
+                a ([1:5, 2:5]):
+                    a: value
+                    ^
+                    a: different_value
+                    ^
+                b ([3:5, 4:5, 5:5]):
+                    b: value
+                    ^
+                    b: different_value
+                    ^
+                    b: different_different_value
+                    ^
+                """)
+        }
+    }
 }
 
 extension YamlErrorTests {